New autonomous test using OOP
